requirements.txt
appdirs==1.4.4
astroid==2.5.6
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.6
beautifulsoup4==4.9.3
click==8.0.1
colorama==0.4.4
Django==2.2.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-forms-bootstrap==3.1.0
fissix==20.8.0
flake8==3.9.2
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
moreorless==0.4.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.8.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
rope==0.19.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
volatile==2.1.0
wrapt==1.12.1
